requirements.txt
asgiref==3.11.1
backports-datetime-fromisoformat==2.0.3
certifi==2026.2.25
charset-normalizer==3.4.7
Django==5.2.13
django-phonenumber-field==8.4.0
django-unfold==0.5.0
djangorestframework==3.17.1
environs==15.0.1
et_xmlfile==2.0.0
gunicorn==25.3.0
idna==3.11
marshmallow==4.3.0
mysqlclient==2.2.8
openpyxl==3.1.5
packaging==26.0
phonenumbers==9.0.27
pillow==12.2.0
python-dotenv==1.2.2
requests==2.33.1
sqlparse==0.5.5
typing_extensions==4.15.0
urllib3==2.6.3
whitenoise==6.12.0
